Fifth, refined working version of Contribution Tree tab.
